+++ src/plugins/pouchdb/pouch-statics.ts
@@ -5,6 +5,7 @@ import {
 import { newRxError } from '../../rx-error';
 
 import {
+    getPouchIndexDesignDocNameByIndex,
     pouchHash,
     pouchSwapPrimaryToId,
     POUCH_HASH_KEY,
@@ -23,6 +24,7 @@ import type {
 } from '../../types';
 import { getPrimaryFieldOfPrimaryKey } from '../../rx-schema';
 import { overwritable } from '../../overwritable';
+import { isMaybeReadonlyArray } from '../../util';
 
 export const RxStoragePouchStatics: RxStorageStatics = {
 
@@ -154,8 +156,11 @@ export function preparePouchDbQuery<RxDocType>(
     if (query.sort) {
         query.sort.forEach(sortPart => {
             const key = Object.keys(sortPart)[0];
-            const comparisonOperators = ['$gt', '$gte', '$lt', '$lte'];
-            const keyUsed = query.selector[key] && Object.keys(query.selector[key]).some(op => comparisonOperators.includes(op)) || false;
+            const comparisonOperators = ['$gt', '$gte', '$lt', '$lte', '$eq'];
+            const keyUsed = query.selector[key] &&
+                Object.keys(query.selector[key]).some(op => comparisonOperators.includes(op))
+                || false; // TODO why we need this '|| false' ?
+
             if (!keyUsed) {
                 const schemaObj = getSchemaByObjectPath(schema, key);
                 if (!schemaObj) {
@@ -229,29 +234,26 @@ export function preparePouchDbQuery<RxDocType>(
         }
     });
 
-    query.selector = primarySwapPouchDbQuerySelector(query.selector, primaryKey);
-
     /**
-     * To ensure a deterministic sorting,
-     * we have to ensure the primary key is always part
-     * of the sort query.
-
-    * TODO This should be done but will not work with pouchdb
-     * because it will throw
-     * 'Cannot sort on field(s) "key" when using the default index'
-     * So we likely have to modify the indexes so that this works. 
+     * Set use_index
+     * @link https://pouchdb.com/guides/mango-queries.html#use_index
      */
-    /*
-    if (!mutateableQuery.sort) {
-        mutateableQuery.sort = [{ [this.primaryPath]: 'asc' }] as any;
-    } else {
-        const isPrimaryInSort = mutateableQuery.sort
-            .find(p => firstPropertyNameOfObject(p) === this.primaryPath);
-        if (!isPrimaryInSort) {
-            mutateableQuery.sort.push({ [this.primaryPath]: 'asc' } as any);
-        }
+    if (mutateableQuery.index) {
+        const indexMaybeArray = mutateableQuery.index;
+        let indexArray: string[] = isMaybeReadonlyArray(indexMaybeArray) ? indexMaybeArray : [indexMaybeArray];
+        indexArray = indexArray.map(str => {
+            if (str === primaryKey) {
+                return '_id';
+            } else {
+                return str;
+            }
+        });
+        const indexName = getPouchIndexDesignDocNameByIndex(indexArray);
+        delete mutateableQuery.index;
+        (mutateableQuery as any).use_index = indexName;
     }
-    */
+
+    query.selector = primarySwapPouchDbQuerySelector(query.selector, primaryKey);
 
     return query;
 }