+++ test/unit/rx-collection.test.ts
@@ -97,11 +97,13 @@ config.parallel('rx-collection.test.js', () => {
                     assert.strictEqual(indexes.indexes.length, 2);
                     const lastIndexDefFields = indexes.indexes[1].def.fields;
                     assert.deepStrictEqual(
-                        lastIndexDefFields, [{
-                            'age': 'asc'
-                        }, {
-                            'passportCountry': 'asc'
-                        }]
+                        lastIndexDefFields,
+                        [
+                            { 'age': 'asc' },
+                            { 'passportCountry': 'asc' },
+                            // the primaryKey index will always be added by RxDB
+                            { _id: 'asc' }
+                        ]
                     );
                     db.destroy();
                 });
@@ -120,11 +122,12 @@ config.parallel('rx-collection.test.js', () => {
                     assert.strictEqual(indexes.indexes.length, 2);
                     const lastIndexDefFields = indexes.indexes[1].def.fields;
                     assert.deepStrictEqual(
-                        lastIndexDefFields, [{
-                            'age': 'asc'
-                        }, {
-                            '|a': 'asc'
-                        }]
+                        lastIndexDefFields, [
+                        { 'age': 'asc' },
+                        { '|a': 'asc' },
+                        // the primaryKey index will always be added by RxDB
+                        { _id: 'asc' }
+                    ]
                     );
                     db.destroy();
                 });
@@ -699,9 +702,10 @@ config.parallel('rx-collection.test.js', () => {
                                     '$gt': 0
                                 }
                             },
-                            sort: [{
-                                'other.age': 'asc'
-                            }]
+                            sort: [
+                                { 'other.age': 'asc' },
+                                { _id: 'asc' }
+                            ]
                         });
                         assert.strictEqual(all.docs.length, 10);