+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### X.X.X (comming soon)
 
+### 12.0.0 (comming soon) BREAKING
+
+- All indexes that do not contain the primaryKey, get the primary key added.
+- You can now set a custom index when doing a query.
+
 - Unified the replication primitives and the GraphQL replication plugin.
 
 ### 11.6.0 (4 February 2022)